+++ packages/db/tests/models/test-granule-model.js
@@ -2,43 +2,27 @@ const test = require('ava');
 const cryptoRandomString = require('crypto-random-string');
 
 const {
-  localStackConnectionEnv,
-  getKnexClient,
   CollectionPgModel,
   ExecutionPgModel,
   GranulePgModel,
-} = require('../../dist');
-const {
+  generateLocalTestDb,
+  destroyLocalTestDb,
   fakeCollectionRecordFactory,
   fakeExecutionRecordFactory,
   fakeGranuleRecordFactory,
-} = require('../../dist/test-utils');
+} = require('../../dist');
 
 const { migrationDir } = require('../../../../lambdas/db-migration');
 
 const testDbName = `granule_${cryptoRandomString({ length: 10 })}`;
-const testDbUser = 'postgres';
 
 test.before(async (t) => {
-  t.context.knexAdmin = await getKnexClient({
-    env: {
-      ...localStackConnectionEnv,
-      migrationDir,
-    },
-  });
-  await t.context.knexAdmin.raw(`create database "${testDbName}";`);
-  await t.context.knexAdmin.raw(`grant all privileges on database "${testDbName}" to "${testDbUser}"`);
-
-  t.context.knex = await getKnexClient({
-    env: {
-      ...localStackConnectionEnv,
-      PG_DATABASE: testDbName,
-      migrationDir,
-    },
-  });
-
-  // create tables
-  await t.context.knex.migrate.latest();
+  const { knexAdmin, knex } = await generateLocalTestDb(
+    testDbName,
+    migrationDir
+  );
+  t.context.knexAdmin = knexAdmin;
+  t.context.knex = knex;
 
   t.context.granulePgModel = new GranulePgModel();
 
@@ -59,9 +43,10 @@ test.before(async (t) => {
 });
 
 test.after.always(async (t) => {
-  await t.context.knex.destroy();
-  await t.context.knexAdmin.raw(`drop database if exists "${testDbName}"`);
-  await t.context.knexAdmin.destroy();
+  await destroyLocalTestDb({
+    ...t.context,
+    testDbName,
+  });
 });
 
 test('GranulePgModel.upsert() creates a new running granule', async (t) => {
@@ -75,6 +60,7 @@ test('GranulePgModel.upsert() creates a new running granule', async (t) => {
   const granule = fakeGranuleRecordFactory({
     collection_cumulus_id: collectionCumulusId,
     execution_cumulus_id: executionCumulusId,
+    status: 'running',
   });
 
   await granulePgModel.upsert(knex, granule);
@@ -85,7 +71,7 @@ test('GranulePgModel.upsert() creates a new running granule', async (t) => {
   );
 });
 
-test('GranulePgModel.upsert() will overwrite allowed fields of a running granule for different execution', async (t) => {
+test('GranulePgModel.upsert() will overwrite allowed fields of a running granule for a different execution', async (t) => {
   const {
     knex,
     executionPgModel,