+++ CHANGELOG.md
@@ -79,6 +79,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     - `@cumulus/message/PDRs/getMessagePdrStats`
     - `@cumulus/message/PDRs/getPdrPercentCompletion`
     - `@cumulus/message/workflows/getWorkflowDuration`
+- **CUMULUS-2199**
+  - Added `translateApiRuleToPostgresRule` to `@cumulus/db` to translate API Rule to conform to Postgres Rule definition.
 
 ### Changed
 
@@ -86,6 +88,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Updated Elasticsearch mappings to support case-insensitive search
 
 - **CUMULUS-2185** - RDS Migration Epic
+  - **CUMULUS-2199**
+    - Updated Rules endpoint to write rules to core PostgreSQL database in addition to DynamoDB and to delete rules from the PostgreSQL database in addition to DynamoDB.
+    - Updated `create` in Rules Model to take in optional `createdAt` parameter which sets the value of createdAt if not specified during function call.
   - **CUMULUS-2189**
     - Updated Provider endpoint logic to write providers in parallel to Core
       PostgreSQL database